fix: CREATE OR REPLACE PAGE reuses existing UUID, int32→int64 for BSON
Root cause fix for Studio Pro crash:
- CREATE OR REPLACE PAGE now reuses the existing page's UUID and calls
  UpdatePage instead of DeletePage+CreatePage. This produces a git
  "Modified" diff instead of "Deleted+Added", preventing Studio Pro's
  RevStatusCache from crashing when parsing deleted mxunit files.

BSON int32→int64:
- All property values in writer serialization changed from int32 to int64
  to match Studio Pro's expected BSON types. Array version markers
  (int32(2), int32(3)) preserved as int32.

diag --check-units:
- New command to detect orphan units and stale mxunit files
- --fix flag auto-removes stale files

+++ mdl/executor/widget_engine.go
@@ -978,7 +978,7 @@ func createClientTemplateBSONWithParams(text string, entityContext string) bson.
 				{Key: "$Type", Value: "Forms$FormattingInfo"},
 				{Key: "CustomDateFormat", Value: ""},
 				{Key: "DateFormat", Value: "Date"},
-				{Key: "DecimalPrecision", Value: int32(2)},
+				{Key: "DecimalPrecision", Value: int64(2)},
 				{Key: "EnumFormat", Value: "Text"},
 				{Key: "GroupDigits", Value: false},
 				{Key: "TimeFormat", Value: "HoursMinutes"},
